+++ src/main/webapp/app/pages/companyPage/CompanyPage.tsx
@@ -10,7 +10,6 @@ import {
   LicenseModel,
   LicenseStatus,
   PAGE_ROUTE,
-  PAGE_TITLE,
   REDIRECT_TIMEOUT_MILLISECONDS,
 } from 'app/config/constants';
 import { Alert, Button, Col, Row } from 'react-bootstrap';
@@ -55,7 +54,6 @@ import { UsageToggleGroup } from 'app/pages/usageAnalysisPage/UsageToggleGroup';
 import OncoKBTable from 'app/components/oncokbTable/OncoKBTable';
 import {
   getUsageTableColumnDefinition,
-  ToggleValue,
   UsageTableColumnKey,
 } from 'app/pages/usageAnalysisPage/UsageAnalysisPage';
 import autobind from 'autobind-decorator';
@@ -71,6 +69,7 @@ import { DownloadButton } from 'app/components/downloadButton/DownloadButton';
 import { RouterStore } from 'mobx-react-router';
 import { TEXT_VAL } from 'app/shared/utils/FormValidationUtils';
 import { Helmet } from 'react-helmet-async';
+import { ToggleValue } from '../usageAnalysisPage/usage-analysis-utils';
 
 interface MatchParams {
   id: string;
