+++ src/main/webapp/app/pages/usageAnalysisPage/UserUsageDetailsTable.tsx
@@ -7,7 +7,6 @@ import React from 'react';
 import { Row } from 'react-bootstrap';
 import {
   getUsageTableColumnDefinition,
-  ToggleValue,
   UsageTableColumnKey,
 } from 'app/pages/usageAnalysisPage/UsageAnalysisPage';
 import {
@@ -32,6 +31,7 @@ import {
   TimeGroupedUsageRecords,
   UsageRecord,
   isPrivateResource,
+  ToggleValue,
 } from './usage-analysis-utils';
 
 type IUserUsageDetailsTable = {
